% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{A build classifier function}
\usage{
build(trainData, className = NA, maxRules = 10000)
}
\arguments{
\item{trainData}{data.frame with train data}

\item{className}{column with target class - default is the last column}

\item{maxRules}{maximum rules in the build model}
}
\value{
data.frame with rules
}
\description{
A build classifier function
}
\examples{
library("rCBA")
data("iris")

train <- sapply(iris, as.factor)
train <- data.frame(train, check.names=FALSE)

model = rCBA::build(train)

predictions <- rCBA::classification(train, model)
table(predictions)
sum(train$Species==predictions, na.rm=TRUE) / length(predictions)

prunedModel <- rCBA::pruning(train, model)

predictions <- rCBA::classification(train, prunedModel)
table(predictions)
sum(train$Species==predictions, na.rm=TRUE) / length(predictions)
}

