% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build2.R
\name{build2}
\alias{build2}
\title{build2}
\usage{
build2(train, consequent = NULL)
}
\arguments{
\item{train}{data.frame or transactions with training dataıı}

\item{consequent}{filter consequent}
}
\description{
build2
}

