% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFPGrowth.R
\name{buildFPGrowth}
\alias{buildFPGrowth}
\title{FP-Growth based build of the model}
\usage{
buildFPGrowth(train, className = NULL)
}
\arguments{
\item{train}{data.frame or transactions with training data}

\item{className}{filter className}
}
\value{
list with parameters and model as data.frame with rules
}
\description{
FP-Growth based build of the model
}
