% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frameToRules}
\alias{frameToRules}
\title{Conversion of data.frame to rules from arules}
\usage{
frameToRules(model)
}
\arguments{
\item{model}{data.frame with rules}
}
\value{
arules rules representation
}
\description{
Conversion of data.frame to rules from arules
}
\examples{
library("rCBA")
data("iris")

output <- rCBA::build(iris)
model <- output$model

rules <- rCBA::frameToRules(model)
inspect(rules)

}
