% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning}
\alias{pruning}
\title{A Pruning function}
\usage{
pruning(train, rules, method = "m2cba")
}
\arguments{
\item{train}{data.frame with training data}

\item{rules}{data.frame with rules}

\item{method}{pruning method m2cba(default)|m1cba| dcbrcba}
}
\value{
data.frame with pruned rules
}
\description{
A Pruning function
}
\examples{
library("arules")
library("rCBA")

train <- read.csv("./train.csv",header=TRUE) # read data

txns <- as(train,"transactions") # convert
rules <- apriori(txns, parameter = list(confidence = 0.1, support= 0.1, minlen=1, maxlen=5)) # rule mining
rules <- subset( rules, subset = rhs \%pin\% "y=") # filter
rulesFrame <- as(rules,"data.frame") # convert

print(nrow(rulesFrame))
prunedRulesFrame <- pruning(trainData, rulesFrame, method="m2cba") # method=m1cba|m2cba|dcbrcba
print(nrow(prunedRulesFrame))
}

