% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{A build classifier function}
\usage{
build(trainData, className = NA, pruning = TRUE)
}
\arguments{
\item{trainData}{data.frame with train data}

\item{className}{column with target class - default is the last column}

\item{pruning}{performing pruning while building the model}
}
\value{
data.frame with rules
}
\description{
A build classifier function
}
\examples{
library("rCBA")
data("iris")

model <- rCBA::build(iris)

predictions <- rCBA::classification(iris, model)
table(predictions)
sum(iris$Species==predictions, na.rm=TRUE) / length(predictions)

}

