% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{A build classifier function}
\usage{
build(trainData, className = NA, pruning = TRUE, sa = list())
}
\arguments{
\item{trainData}{data.frame with train data}

\item{className}{column with target class - default is the last column}

\item{pruning}{performing pruning while building the model}

\item{sa}{simulated annealing setting. default values list(temp=100.0, alpha=0.05, tabuRuleLength=5, timeout=10)}
}
\value{
list with parameters and model as data.frame with rules
}
\description{
A build classifier function
}
\examples{
library("rCBA")
data("iris")

output <- rCBA::build(iris)
model <- output$model

predictions <- rCBA::classification(iris, model)
table(predictions)
sum(iris$Species==predictions, na.rm=TRUE) / length(predictions)

}

