% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning}
\alias{pruning}
\title{A Pruning function}
\usage{
pruning(train, rules, method = "m2cba")
}
\arguments{
\item{train}{data.frame with training data}

\item{rules}{data.frame with rules}

\item{method}{pruning method m2cba(default)|m1cba|dcbrcba}
}
\value{
data.frame with pruned rules
}
\description{
A Pruning function
}
\examples{
library("arules")
library("rCBA")
data("AdultUCI")
data("Adult")

rules <- apriori(Adult, parameter=list(support=0.1, confidence=0.1), 
appearance = list(rhs=c("sex=Male", "sex=Female"),default="lhs"))
rulesFrame <- as(rules,"data.frame") # convert

print(nrow(rulesFrame))
prunedRulesFrame <- pruning(AdultUCI, rulesFrame, method="m2cba")
print(nrow(prunedRulesFrame))
}

